fix logic

+++ src/k8s/pkg/k8sd/api/cluster_remove.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 
 	apiv1 "github.com/canonical/k8s/api/v1"
+	databaseutil "github.com/canonical/k8s/pkg/k8sd/database/util"
 	"github.com/canonical/k8s/pkg/log"
 	"github.com/canonical/k8s/pkg/utils"
 	"github.com/canonical/k8s/pkg/utils/control"
@@ -72,18 +73,25 @@ func (e *Endpoints) postClusterRemove(s state.State, r *http.Request) response.R
 		return response.SyncResponse(true, nil)
 	}
 
-	client, err := snap.KubernetesClient("")
+	cfg, err := databaseutil.GetClusterConfig(ctx, s)
 	if err != nil {
-		return response.InternalError(fmt.Errorf("failed to create k8s client: %w", err))
-	}
-	if node, err := client.CoreV1().Nodes().Get(ctx, req.Name, metav1.GetOptions{}); err != nil {
-		return NodeUnavailable(fmt.Errorf("node %q is not part of the cluster: %w", req.Name, err))
-	} else if v, ok := node.Labels["k8sd.io/role"]; !ok || v != "worker" {
-		return NodeUnavailable(fmt.Errorf("node %q is missing k8sd.io/role=worker label", req.Name))
+		return response.InternalError(fmt.Errorf("failed to get cluster config: %w", err))
 	}
 
-	if err := client.DeleteNode(ctx, req.Name); err != nil {
-		return response.InternalError(fmt.Errorf("failed to remove k8s node %q: %w", req.Name, err))
+	if cfg.K8sd.GetShouldRemoveK8sNode() {
+		client, err := snap.KubernetesClient("")
+		if err != nil {
+			return response.InternalError(fmt.Errorf("failed to create k8s client: %w", err))
+		}
+		if node, err := client.CoreV1().Nodes().Get(ctx, req.Name, metav1.GetOptions{}); err != nil {
+			return NodeUnavailable(fmt.Errorf("node %q is not part of the cluster: %w", req.Name, err))
+		} else if v, ok := node.Labels["k8sd.io/role"]; !ok || v != "worker" {
+			return NodeUnavailable(fmt.Errorf("node %q is missing k8sd.io/role=worker label", req.Name))
+		}
+
+		if err := client.DeleteNode(ctx, req.Name); err != nil {
+			return response.InternalError(fmt.Errorf("failed to remove k8s node %q: %w", req.Name, err))
+		}
 	}
 
 	return response.SyncResponse(true, nil)
